refactor(objectives): move axis equivalence into discovery
Why:
- A model judgment that only classifies backend-selected label pairs was exposed as candidate-generation behavior on the shared Objective extractor.

What:
- Give axis equivalence its own discovery contract and inject it into deterministic Objective candidate grouping.

Impact:
- Pair selection, all-or-nothing fallback, Objective grouping, lineage, confidence, persistence, and public contracts remain unchanged.

Verification:
- All 143 candidate and extractor tests and 6 collection-build integration tests passed.
