+++ backend/tests/support/objective_extractor.py
@@ -3,12 +3,14 @@ from __future__ import annotations
 from typing import Any
 
 from application.core.document_profiles.schemas import StructuredDocumentProfile
+from application.core.objectives.discovery.axis_equivalence import (
+    StructuredAxisCanonicalizationPlan,
+)
 from application.core.objectives.discovery.signal_reconciliation import (
     StructuredPaperSignalReconciliation,
 )
 from application.core.objectives.discovery.study_window import StructuredPaperSkim
 from application.core.objectives.schemas import (
-    StructuredAxisCanonicalizationPlan,
     StructuredEvidenceExtraction,
     StructuredEvidenceExtractions,
     StructuredEvidenceSelection,
@@ -132,7 +134,7 @@ class FakeObjectiveExtractor:
     ) -> StructuredPaperSignalReconciliation:
         return StructuredPaperSignalReconciliation()
 
-    def canonicalize_research_objective_axes(
+    def classify(
         self,
         payload: dict[str, Any],
     ) -> StructuredAxisCanonicalizationPlan:
